+++ src/core/index.ts
@@ -1,2 +1,5 @@
 export * from './chord';
+export * from './degree';
+export * from './key';
+export * from './notation';
 export * from './pitch';